Update commands and queries for requests, time for lessons, profile, update controllers, add edit-calendar, details-calendar and other

+++ Domain/Commands/DeleteLessonTimeCommand.cs
@@ -1,22 +1,38 @@
+using System.Data;
 using Infra.DatabaseAdapter;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using AutoMapper;
+using Domain.Port.Driving;
 
 namespace Domain.Commands;
 
 public class DeleteLessonTimeCommand : IRequest<bool>
 {
+    public TimeType Type { get; set; }
     public int UpdatedBy { get; set; }
     public int EventId { get; set; }
 
     public class DeleteLessonTimeCommandHandler : BaseMediatrHandler<DeleteLessonTimeCommand, bool>
     {
-        public override async Task<bool> Handle(DeleteLessonTimeCommand request,
-            CancellationToken cancellationToken)
+        public override async Task<bool> Handle(DeleteLessonTimeCommand r, CancellationToken token)
         {
-            var lesson = ApplicationDb.Lessons.Where(x => x.Id == request.EventId).First();
-            ApplicationDb.Remove(lesson);
+            if (r.Type == TimeType.Available)
+            {
+                var lesson = ApplicationDb.AvailableTimes.Where(x => x.Id == r.EventId).FirstOrDefault();
+                if (lesson == null)
+                    throw new ArgumentNullException("Помилка в номері події");
+                ApplicationDb.Remove(lesson);
+            }
+            else
+            {
+                var lesson = ApplicationDb.Lessons.Where(x => x.Id == r.EventId).FirstOrDefault();
+
+                if (lesson == null)
+                    throw new ArgumentNullException("Помилка в номері події");
+                ApplicationDb.Remove(lesson);
+            }
+
             await ApplicationDb.SaveChangesAsync();
             return true;
         }
